Question_2/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
pydantic==2.6.0